// src/generated/models/ServiceOrdersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface ServiceOrdersBase {
  ID?: number;
  Title?: string;
  Customer?: string;
  FormCompletionTime?: string;
  CustomerName?: string;
  CustomerLocation?: string;
  CustomerTenant?: string;
  IsAdhocCustomer?: boolean;
  DriverName?: string;
  VehicleNumber?: string;
  DateOfCollection?: string;
  ReceiptUrl?: string;
  SignatureUrl?: string;
  BeforePhotoUrl?: string;
  AfterPhotoUrl?: string;
  WasteCategory?: string;
  Tonnage?: number;
  Notes?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface ServiceOrdersRead extends ServiceOrdersBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface ServiceOrdersWrite extends ServiceOrdersBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}